more remote_operation to separate dto class
