+++ spec/indocker/launchers/configuration_deployer_spec.rb
@@ -4,9 +4,9 @@ RSpec.describe Indocker::Launchers::ConfigurationDeployer do
   before { setup_indocker(debug: true) }
 
   def build_remote_operation
-    thread = Thread.new() do
-    end
-    Indocker::Launchers::ConfigurationDeployer::RemoteOperation.new(thread, :external, :indocker_sync)
+    thread = Thread.new() {}
+
+    Indocker::Launchers::DTO::RemoteOperationDTO.new(thread, :external, :indocker_sync)
   end
 
   subject { Indocker::Launchers::ConfigurationDeployer.new(
@@ -18,7 +18,7 @@ RSpec.describe Indocker::Launchers::ConfigurationDeployer do
     deployment_policy = build_deployment_policy({
       containers: [:good_container]
     })
-    
+
     expect(subject).to receive(:compile_image).once.and_return([build_remote_operation])
     expect(subject).to receive(:deploy_container).once.and_return([build_remote_operation])
     expect(subject).to receive(:sync_indocker).once.and_return([build_remote_operation])
